% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_ethnicity_population_deprivation.R,
%   R/facet_ethnicity_sources_weighted.R
\name{facet_ethnicity_sources_weighted}
\alias{facet_ethnicity_sources_weighted}
\title{Bind ethnicity data to the prawn and create graphs based on it}
\usage{
facet_ethnicity_sources_weighted(prawn_path, pollutant, year)

facet_ethnicity_sources_weighted(prawn_path, pollutant, year)
}
\arguments{
\item{prawn_path}{The filepath for the prawn CSV that is to be used.}

\item{pollutant}{The pollutant being investigated, used in graph titles}

\item{year}{The year being investigated, used in graph titles}
}
\description{
Bind ethnicity data to the prawn and create graphs based on it

Bind ethnicity data to the prawn and create graphs based on it
}
\examples{
cartesian_ethnicity_groups_src(
  prawn_path="PRAWN.csv",
  pollutant="NOx",
  year=2019)
cartesian_ethnicity_groups_src(
  prawn_path="PRAWN.csv",
  pollutant="NOx",
  year=2019)
}
\keyword{faceted,}
\keyword{sources}
